pyproject.toml
[project]
name = "mori"
version = "0.1.0"
description = "基于 AgentScope 的虚拟 AI 女友系统"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agentscope>=1.0.8",
    "gradio>=4.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "openai>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/moria"
Documentation = "https://github.com/yourusername/moria/blob/main/README.md"
Repository = "https://github.com/yourusername/moria"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.hatch.build.targets.wheel]
packages = ["mori"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
]
